refactor: move related SingletonEditor code from Main to the right module
