+++ src/components/Main.tsx
@@ -2,64 +2,12 @@ import { Link, Outlet } from "react-router-dom";
 import { useAppServices } from "./AppServicesContext";
 import logo from "./logo.svg";
 import "./Main.css";
-import {
-  EditorState,
-  emptyDesign,
-  ISingletonDesignContext,
-  SingletonDesignContext,
-  SingletonEditor,
-} from "./SingletonEditor";
-import { Design, HtmlExport } from "react-email-editor";
-import { useEffect, useState } from "react";
+import { SingletonEditorProvider } from "./SingletonEditor";
 
 export function Main() {
   const {
     appConfiguration: { loginPageUrl },
   } = useAppServices();
-
-  const [design, setDesign] = useState<Design | undefined>();
-  const hidden = !design;
-  const [editorState, setEditorState] = useState<EditorState>({
-    unlayer: undefined,
-    isLoaded: false,
-  });
-
-  const getHtml = () => {
-    if (!editorState.isLoaded) {
-      return Promise.resolve("");
-    }
-    return new Promise<string>((resolve) => {
-      editorState.unlayer.exportHtml((htmlExport: HtmlExport) => {
-        resolve(htmlExport.html);
-      });
-    });
-  };
-
-  const getDesign = () => {
-    if (!editorState.isLoaded) {
-      return Promise.resolve(emptyDesign);
-    }
-    return new Promise<Design>((resolve) => {
-      editorState.unlayer.exportHtml((htmlExport: HtmlExport) => {
-        resolve(htmlExport.design);
-      });
-    });
-  };
-
-  useEffect(() => {
-    if (editorState.isLoaded) {
-      editorState.unlayer.loadDesign(design || emptyDesign);
-    }
-  }, [design, editorState]);
-
-  const defaultContext: ISingletonDesignContext = {
-    hidden,
-    setDesign,
-    unsetDesign: () => setDesign(undefined),
-    setEditorState,
-    getDesign,
-    getHtml,
-  };
 
   return (
     <div className="App">
@@ -80,10 +28,9 @@ export function Main() {
           </div>
         </nav>
       </header>
-      <SingletonDesignContext.Provider value={defaultContext}>
+      <SingletonEditorProvider>
         <Outlet />
-        <SingletonEditor />
-      </SingletonDesignContext.Provider>
+      </SingletonEditorProvider>
     </div>
   );
 }
