+++ src/components/SingletonEditor.tsx
@@ -1,6 +1,6 @@
-import { createContext, useContext } from "react";
+import { createContext, useContext, useEffect, useState } from "react";
 import { Editor } from "./Editor";
-import EmailEditor, { Design } from "react-email-editor";
+import EmailEditor, { Design, HtmlExport } from "react-email-editor";
 
 export type EditorState =
   | { isLoaded: false; unlayer: undefined }
@@ -37,3 +37,60 @@ export const SingletonEditor = (props: any) => {
 };
 
 export const useSingletonEditor = () => useContext(SingletonDesignContext);
+
+export const SingletonEditorProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const [design, setDesign] = useState<Design | undefined>();
+  const hidden = !design;
+  const [editorState, setEditorState] = useState<EditorState>({
+    unlayer: undefined,
+    isLoaded: false,
+  });
+
+  const getHtml = () => {
+    if (!editorState.isLoaded) {
+      return Promise.resolve("");
+    }
+    return new Promise<string>((resolve) => {
+      editorState.unlayer.exportHtml((htmlExport: HtmlExport) => {
+        resolve(htmlExport.html);
+      });
+    });
+  };
+
+  const getDesign = () => {
+    if (!editorState.isLoaded) {
+      return Promise.resolve(emptyDesign);
+    }
+    return new Promise<Design>((resolve) => {
+      editorState.unlayer.exportHtml((htmlExport: HtmlExport) => {
+        resolve(htmlExport.design);
+      });
+    });
+  };
+
+  useEffect(() => {
+    if (editorState.isLoaded) {
+      editorState.unlayer.loadDesign(design || emptyDesign);
+    }
+  }, [design, editorState]);
+
+  const defaultContext: ISingletonDesignContext = {
+    hidden,
+    setDesign,
+    unsetDesign: () => setDesign(undefined),
+    setEditorState,
+    getDesign,
+    getHtml,
+  };
+
+  return (
+    <SingletonDesignContext.Provider value={defaultContext}>
+      {children}
+      <SingletonEditor />
+    </SingletonDesignContext.Provider>
+  );
+};
